[BEAM-7372] cleanup codes for py2 from apache_beam/transforms (#14544)
