+++ sdks/python/apache_beam/transforms/__init__.py
@@ -18,7 +18,6 @@
 """PTransform and descendants."""
 
 # pylint: disable=wildcard-import
-from __future__ import absolute_import
 
 from apache_beam.transforms import combiners
 from apache_beam.transforms.core import *
