+++ sdks/python/apache_beam/transforms/dataflow_distribution_counter_test.py
@@ -16,8 +16,6 @@ otherwise, test on pure python module
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import unittest
 
 from mock import Mock
