+++ sdks/python/apache_beam/transforms/environments.py
@@ -21,8 +21,6 @@ For internal use only. No backwards compatibility guarantees."""
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import json
 import logging
 import sys
