+++ sdks/python/apache_beam/transforms/environments_test.py
@@ -20,8 +20,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import logging
 import unittest
