+++ sdks/python/apache_beam/transforms/external_it_test.py
@@ -19,8 +19,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import unittest
 
 from nose.plugins.attrib import attr
