+++ sdks/python/apache_beam/transforms/periodicsequence.py
@@ -15,9 +15,6 @@
 # limitations under the License.
 #
 
-from __future__ import absolute_import
-from __future__ import division
-
 import math
 import time
