+++ sdks/python/apache_beam/transforms/sideinputs_test.py
@@ -19,8 +19,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import itertools
 import logging
 import unittest
