+++ sdks/python/apache_beam/transforms/timeutil.py
@@ -19,13 +19,8 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 from abc import ABCMeta
 from abc import abstractmethod
-from builtins import object
-
-from future.utils import with_metaclass
 
 from apache_beam.portability.api import beam_runner_api_pb2
 
@@ -63,7 +58,7 @@ class TimeDomain(object):
     return TimeDomain.from_string(domain) == TimeDomain.WATERMARK
 
 
-class TimestampCombinerImpl(with_metaclass(ABCMeta, object)):  # type: ignore[misc]
+class TimestampCombinerImpl(metaclass=ABCMeta):
   """Implementation of TimestampCombiner."""
   @abstractmethod
   def assign_output_time(self, window, input_timestamp):
@@ -88,7 +83,7 @@ class TimestampCombinerImpl(with_metaclass(ABCMeta, object)):  # type: ignore[mi
     return self.combine_all(merging_timestamps)
 
 
-class DependsOnlyOnWindow(with_metaclass(ABCMeta, TimestampCombinerImpl)):  # type: ignore[misc]
+class DependsOnlyOnWindow(TimestampCombinerImpl, metaclass=ABCMeta):
   """TimestampCombinerImpl that only depends on the window."""
   def merge(self, result_window, unused_merging_timestamps):
     # Since we know that the result only depends on the window, we can ignore
