+++ sdks/python/apache_beam/transforms/transforms_keyword_only_args_test.py
@@ -19,8 +19,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import logging
 import unittest
