+++ sdks/python/apache_beam/transforms/userstate.py
@@ -23,11 +23,8 @@ Experimental; no backwards-compatibility guarantees.
 # pytype: skip-file
 # mypy: disallow-untyped-defs
 
-from __future__ import absolute_import
-
 import collections
 import types
-from builtins import object
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import Callable
