+++ sdks/python/apache_beam/transforms/util.py
@@ -20,9 +20,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-from __future__ import division
-
 import collections
 import contextlib
 import random
@@ -30,10 +27,6 @@ import re
 import threading
 import time
 import uuid
-from builtins import filter
-from builtins import object
-from builtins import range
-from builtins import zip
 from typing import TYPE_CHECKING
 from typing import Any
 from typing import Iterable
@@ -41,8 +34,6 @@ from typing import List
 from typing import Tuple
 from typing import TypeVar
 from typing import Union
-
-from future.utils import itervalues
 
 from apache_beam import coders
 from apache_beam import typehints
@@ -162,7 +153,7 @@ class CoGroupByKey(PTransform):
   def _extract_input_pvalues(self, pvalueish):
     try:
       # If this works, it's a dict.
-      return pvalueish, tuple(itervalues(pvalueish))
+      return pvalueish, tuple(pvalueish.values())
     except AttributeError:
       pcolls = tuple(pvalueish)
       return pcolls, pcolls
